Split up some filed and clean up a bit. Passing 15/30
